Move logrus wrapper to ocis-pkg
To make it more widely usable
